+++ Basic/ClientDriven/Assets/Scripts/ClientPlayerMove.cs
@@ -1,6 +1,9 @@
+using System;
+using Cinemachine;
+using StarterAssets;
 using Unity.Netcode;
-using Unity.Netcode.Samples;
 using UnityEngine;
+using UnityEngine.InputSystem;
 
 /// <summary>
 /// Assumes client authority
@@ -9,24 +12,40 @@ using UnityEngine;
 [DefaultExecutionOrder(1)] // after server component
 public class ClientPlayerMove : NetworkBehaviour
 {
-    private ServerPlayerMove m_Server;
+    [SerializeField]
+    ServerPlayerMove m_ServerPlayerMove;
 
     [SerializeField]
-    private float m_Speed = 5;
+    CharacterController m_CharacterController;
 
     [SerializeField]
-    private float m_RotSpeed = 5;
+    ThirdPersonController m_ThirdPersonController;
 
     [SerializeField]
-    private Camera m_Camera;
+    CapsuleCollider m_CapsuleCollider;
 
-    private CharacterController m_CharacterController;
+    [SerializeField]
+    Transform m_CameraFollow;
 
-    private void Awake()
+    [SerializeField]
+    PlayerInput m_PlayerInput;
+
+    RaycastHit[] m_HitColliders = new RaycastHit[4];
+
+    void Awake()
     {
-        m_Server = GetComponent<ServerPlayerMove>();
         Cursor.lockState = CursorLockMode.Locked;
         Cursor.visible = false;
+        
+        // ThirdPersonController & CharacterController are enabled only on owning clients. Ghost player objects have
+        // these two components disabled, and will enable a CapsuleCollider. Per the CharacterController documentation: 
+        // https://docs.unity3d.com/Manual/CharacterControllers.html, a Character controller can push rigidbody
+        // objects aside while moving but will not be accelerated by incoming collisions. This means that a primitive
+        // CapusleCollider must instead be used for ghost clients to simulate collisions between owning players and 
+        // ghost clients.
+        m_ThirdPersonController.enabled = false;
+        m_CapsuleCollider.enabled = false;
+        m_CharacterController.enabled = false;
     }
 
     public override void OnNetworkSpawn()
@@ -36,58 +55,51 @@ public class ClientPlayerMove : NetworkBehaviour
         enabled = IsClient;
         if (!IsOwner)
         {
-            m_Camera.gameObject.SetActive(false);
             enabled = false;
+            m_CharacterController.enabled = false;
+            m_CapsuleCollider.enabled = true;
             return;
         }
 
-        m_CharacterController = GetComponent<CharacterController>();
-    }
-
-    [ClientRpc]
-    public void SetSpawnClientRpc(Vector3 position, ClientRpcParams clientRpcParams = default)
-    {
-        m_CharacterController.enabled = false;
-        transform.position = position;
+        // player input is only enabled on owning players
+        m_PlayerInput.enabled = true;
+        m_ThirdPersonController.enabled = true;
+        
+        // see the note inside ServerPlayerMove why this step is also necessary for synchronizing initial player
+        // position on owning clients
         m_CharacterController.enabled = true;
-        gameObject.SetActive(true);
+
+        var cinemachineVirtualCamera = FindObjectOfType<CinemachineVirtualCamera>();
+        cinemachineVirtualCamera.Follow = m_CameraFollow;
     }
 
-    // DOC START HERE
-    void FixedUpdate()
+    void OnPickUp()
     {
-        // enabled = false if we're not the owner, so no need to guard the following code if isOwner check
-        // move client authoritative object. Those moves will be replicated to other players using ClientNetworkTransform (taken from Netcode's samples)
-        Vector3 move = Input.GetAxis("Horizontal") * transform.right + Input.GetAxis("Vertical") * transform.forward;
-        m_CharacterController.SimpleMove(move * Time.deltaTime * m_Speed);
-        transform.Rotate(0, (Input.GetAxis("Mouse X")) * Time.deltaTime * m_RotSpeed, 0);
-    }
-    // DOC END HERE
-
-    void Update()
-    {
-        if (Input.GetKeyDown("e"))
+        if (m_ServerPlayerMove.isObjectPickedUp.Value)
         {
-            if (m_Server.ObjPickedUp.Value)
+            m_ServerPlayerMove.DropObjectServerRpc();
+        }
+        else
+        {
+            // detect nearby ingredients
+            var hits = Physics.BoxCastNonAlloc(transform.position,
+                Vector3.one,
+                transform.forward,
+                m_HitColliders,
+                Quaternion.identity,
+                1f,
+                LayerMask.GetMask(new[] { "PickupItems" }),
+                QueryTriggerInteraction.Ignore);
+            if (hits > 0)
             {
-                m_Server.DropObjServerRpc();
-            }
-            else
-            {
-                // todo use non-alloc, don't do the below at home
-                // detect nearby ingredients
-                var hit = Physics.OverlapSphere(transform.position, 5, LayerMask.GetMask(new[] {"PickupItems"}), QueryTriggerInteraction.Ignore);
-                if (hit.Length > 0)
+                var ingredient = m_HitColliders[0].collider.gameObject.GetComponent<ServerIngredient>();
+                if (ingredient != null)
                 {
-                    var ingredient = hit[0].gameObject.GetComponent<ServerIngredient>();
-                    if (ingredient != null)
-                    {
-                        var netObj = ingredient.NetworkObjectId;
-                        // Netcode is a server driven SDK. Shared objects like ingredients need to be interacted with using ServerRPCs. Therefore, there
-                        // will be a delay between the button press and the reparenting.
-                        // This delay could be hidden with some animations/sounds/VFX that would be triggered here.
-                        m_Server.PickupObjServerRpc(netObj);
-                    }
+                    var netObj = ingredient.NetworkObjectId;
+                    // Netcode is a server driven SDK. Shared objects like ingredients need to be interacted with using ServerRPCs. Therefore, there
+                    // will be a delay between the button press and the reparenting.
+                    // This delay could be hidden with some animations/sounds/VFX that would be triggered here.
+                    m_ServerPlayerMove.PickupObjectServerRpc(netObj);
                 }
             }
         }