+++ Basic/ClientDriven/Assets/Scripts/ServerIngredient.cs
@@ -1,3 +1,4 @@
+using System;
 using Unity.Netcode;
 using UnityEngine;
 
@@ -7,14 +8,18 @@ public class ServerIngredient : ServerObjectWithIngredientType
 
 public enum IngredientType
 {
-    red,
-    blue,
-    purple,
-    max // should be always last
+    Red,
+    Blue,
+    Purple,
+    MAX // should be always last
 }
 
 public class ServerObjectWithIngredientType : NetworkBehaviour
 {
+    public NetworkVariable<IngredientType> currentIngredientType;
+
+    public event Action ingredientDespawned;
+    
     public override void OnNetworkSpawn()
     {
         base.OnNetworkSpawn();
@@ -25,6 +30,8 @@ public class ServerObjectWithIngredientType : NetworkBehaviour
         }
     }
 
-    [SerializeField]
-    public NetworkVariable<IngredientType> CurrentIngredientType;
+    public override void OnNetworkDespawn()
+    {
+        ingredientDespawned?.Invoke();
+    }
 }