+++ Basic/Invaders/Assets/Scripts/SuperEnemyMovement.cs
@@ -1,23 +1,23 @@
 ﻿using Unity.Netcode;
 using UnityEngine;
 
-public class SuperEnemyMovement : MonoBehaviour
+public class SuperEnemyMovement : NetworkBehaviour
 {
     private const float k_YBoundary = 14.0f;
 
     // The constant speed at which the Bullet travels
     [Header("Movement Settings")]
     [SerializeField]
-    [Tooltip("The constant speed at which the Sauce moves")]
+    [Tooltip("The constant speed at which the Saucer moves")]
     private float m_MoveSpeed = 3.5f;
 
     private void Update()
     {
-        if (!NetworkManager.Singleton.IsServer) return;
+        if (!IsServer) return;
         
         if (transform.position.x > k_YBoundary)
         {
-            if (NetworkManager.Singleton.IsServer) Destroy(gameObject);
+            if (IsServer) NetworkObject.Despawn();
             return;
         }